requirements.txt
anthropic>=0.18.0
openai>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
markdown>=3.5.0
feedparser>=6.0.0
PyMySQL>=1.1.0
psycopg2-binary>=2.9.9
